simplify navigation logic
