+++ app/src/Navigation/home.ts
@@ -1,0 +1,27 @@
+import { NativeStackNavigationOptions } from '@react-navigation/native-stack';
+
+import HomeNavBar from '../components/HomeNavBar';
+import DisclaimerScreen from '../screens/DisclaimerScreen';
+import HomeScreen from '../screens/HomeScreen';
+import { black } from '../utils/colors';
+
+const homeScreens = {
+  Home: {
+    screen: HomeScreen,
+    options: {
+      title: 'Self',
+      header: HomeNavBar,
+      navigationBarColor: black,
+      presentation: 'card',
+    } as NativeStackNavigationOptions,
+  },
+  Disclaimer: {
+    screen: DisclaimerScreen,
+    options: {
+      title: 'Disclaimer',
+      headerShown: false,
+    } as NativeStackNavigationOptions,
+  },
+};
+
+export default homeScreens;
