+++ app/src/Navigation/index.tsx
@@ -1,0 +1,94 @@
+import React from 'react';
+import 'react-native-gesture-handler';
+import { GestureHandlerRootView } from 'react-native-gesture-handler';
+
+import {
+  StaticParamList,
+  createNavigationContainerRef,
+  createStaticNavigation,
+} from '@react-navigation/native';
+import { createNativeStackNavigator } from '@react-navigation/native-stack';
+
+import DefaultNavBar from '../components/DefaultNavBar';
+import AppLayout from '../layouts/AppLayout';
+import { useApp } from '../stores/appProvider';
+import { useProofInfo } from '../stores/proofProvider';
+import analytics from '../utils/analytics';
+import { white } from '../utils/colors';
+import { setupUniversalLinkListenerInNavigation } from '../utils/qrCodeNew';
+import accountScreens from './account';
+import homeScreens from './home';
+import passportScreens from './passport';
+import proveScreens from './prove';
+import settingsScreens from './settings';
+import staticScreens from './static';
+
+const AppNavigation = createNativeStackNavigator({
+  initialRouteName: 'Splash',
+  orientation: 'portrait_up',
+  screenOptions: {
+    header: DefaultNavBar,
+    navigationBarColor: white,
+  },
+  layout: AppLayout,
+  screens: {
+    ...staticScreens,
+    ...passportScreens,
+    ...homeScreens,
+    ...proveScreens,
+    ...accountScreens,
+    ...settingsScreens,
+  },
+});
+
+export type RootStackParamList = StaticParamList<typeof AppNavigation>;
+
+declare global {
+  namespace ReactNavigation {
+    interface RootParamList extends RootStackParamList {}
+  }
+}
+
+// Create a ref that we can use to access the navigation state
+export const navigationRef = createNavigationContainerRef();
+
+const { trackScreenView } = analytics();
+
+const Navigation = createStaticNavigation(AppNavigation);
+const NavigationWithTracking = () => {
+  const trackScreen = () => {
+    const currentRoute = navigationRef.getCurrentRoute();
+    if (currentRoute) {
+      console.log(`Screen View: ${currentRoute.name}`);
+      trackScreenView(`${currentRoute.name}`, {
+        screenName: currentRoute.name,
+      });
+    }
+  };
+
+  // Add these hooks to get access to the necessary functions
+  const { setSelectedApp, cleanSelfApp } = useProofInfo();
+  const { startAppListener } = useApp();
+
+  // Setup universal link handling at the navigation level
+  React.useEffect(() => {
+    const cleanup = setupUniversalLinkListenerInNavigation(
+      navigationRef,
+      setSelectedApp,
+      cleanSelfApp,
+      startAppListener,
+    );
+
+    return () => {
+      cleanup();
+    };
+  }, [setSelectedApp, cleanSelfApp, startAppListener]);
+
+  return (
+    <GestureHandlerRootView>
+      <Navigation ref={navigationRef} onStateChange={trackScreen} />
+    </GestureHandlerRootView>
+  );
+};
+
+export default NavigationWithTracking;
