+++ app/src/Navigation/settings.ts
@@ -1,0 +1,69 @@
+import { NativeStackNavigationOptions } from '@react-navigation/native-stack';
+
+import CloudBackupScreen from '../screens/Settings/CloudBackupScreen';
+import DevSettingsScreen from '../screens/Settings/DevSettingsScreen';
+import PassportDataInfoScreen from '../screens/Settings/PassportDataInfoScreen';
+import ShowRecoveryPhraseScreen from '../screens/Settings/ShowRecoveryPhraseScreen';
+import SettingsScreen from '../screens/SettingsScreen';
+import { black, slate300, white } from '../utils/colors';
+
+const settingsScreens = {
+  Settings: {
+    screen: SettingsScreen,
+    options: {
+      animation: 'slide_from_bottom',
+      title: 'Settings',
+      headerStyle: {
+        backgroundColor: white,
+      },
+      headerTitleStyle: {
+        color: black,
+      },
+      navigationBarColor: black,
+    } as NativeStackNavigationOptions,
+    config: {
+      screens: {},
+    },
+  },
+  ShowRecoveryPhrase: {
+    screen: ShowRecoveryPhraseScreen,
+    options: {
+      title: 'Recovery Phrase',
+      headerStyle: {
+        backgroundColor: white,
+      },
+    } as NativeStackNavigationOptions,
+  },
+  PassportDataInfo: {
+    screen: PassportDataInfoScreen,
+    options: {
+      title: 'Passport Data Info',
+      headerStyle: {
+        backgroundColor: white,
+      },
+    } as NativeStackNavigationOptions,
+  },
+  DevSettings: {
+    screen: DevSettingsScreen,
+    options: {
+      title: 'Developer Settings',
+      headerStyle: {
+        backgroundColor: white,
+      },
+    } as NativeStackNavigationOptions,
+  },
+  CloudBackupSettings: {
+    screen: CloudBackupScreen,
+    options: {
+      title: 'Cloud backup',
+      headerStyle: {
+        backgroundColor: black,
+      },
+      headerTitleStyle: {
+        color: slate300,
+      },
+    } as NativeStackNavigationOptions,
+  },
+};
+
+export default settingsScreens;
